added case assignments
