+++ queues/urls.py
@@ -10,4 +10,6 @@ urlpatterns = [
     path('add', views.AddQueue.as_view(), name='add'),
     # ex: /queues/43a88949-5db9-4334-b0cc-044e91827451/ - Go to a queues's detail
     path('<uuid:pk>/', views.EditQueue.as_view(), name='edit'),
+    # ex: /queues/43a88949-5db9-4334-b0cc-044e91827451/case-assignments/
+    path('<uuid:pk>/case-assignments/', views.CaseAssignments.as_view(), name='case_assignments'),
 ]
